Moved Resource Loader to exchange-connector module
