+++ src/exchange-connector/index.ts
@@ -18,4 +18,5 @@ export {
 export { groupByCategory } from "./exchangeTools";
 export { extractFiles } from "./exchangeDirectoryParser";
 
+export { CommerceStandardsLoader } from "./resourceLoader";
 export { RestApi } from "./exchangeTypes";
